Update blueprint.yaml
Previous the flash only works 1 time or 15 now it works with a loop. As soon as the status is SC or VSC it will continues flashing (on, 1s pause, off, 1s pause) untill the track_status changes back to a "fixed" flag. The flash option is *not* used because not every device is supporting this option.